fix(global): installing scss
