+++ App.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import Prelogin from './src/components/prelogin';
-import login from './src/components/login';
+import Prelogin from './src/components/pages/prelogin';
+import Login from './src/components/pages/login';
+import Register from './src/components/pages/register';
 
 const Stack = createStackNavigator();
 
@@ -10,8 +11,9 @@ function App() {
   return (
     <NavigationContainer>
       <Stack.Navigator initialRouteName="PreLogin">
-        <Stack.Screen name="PreLogin" component={Prelogin} />
-        <Stack.Screen name="Login" component={login} />
+        <Stack.Screen name="Prelogin" component={Prelogin} />
+        <Stack.Screen name="Login" component={Login} />
+        <Stack.Screen name="Register" component={Register} />
       </Stack.Navigator>
     </NavigationContainer>
   );
