+++ metro.config.js
@@ -1,0 +1,19 @@
+const {
+    getDefaultConfig
+} = require("metro-config");
+
+module.exports = (async () => {
+    const {
+        resolver: {
+            sourceExts
+        }
+    } = await getDefaultConfig();
+    return {
+        transformer: {
+            babelTransformerPath: require.resolve("./transformer.js")
+        },
+        resolver: {
+            sourceExts: [...sourceExts, "scss", "sass"]
+        }
+    };
+})();
