+++ package.json
@@ -18,13 +18,17 @@
     "commitizen": "^4.0.4",
     "expo": "~37.0.3",
     "graphql": "^15.0.0",
+    "postcss": "^7.0.28",
+    "postcss-css-variables": "^0.17.0",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
     "react-native-elements": "^2.0.0",
     "react-native-gesture-handler": "~1.6.0",
+    "react-native-postcss-transformer": "^1.2.4",
     "react-native-reanimated": "~1.7.0",
     "react-native-safe-area-context": "0.7.3",
+    "react-native-sass-transformer": "^1.4.0",
     "react-native-screens": "~2.2.0",
     "react-native-vector-icons": "^6.6.0",
     "react-native-web": "~0.11.7"
