+++ src/components/facebookButton.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import { Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { COLORS } from '../styles/constant'
+import Icon from 'react-native-vector-icons/FontAwesome';
+
+function FacebookButton({ title }) {
+    return (
+        <TouchableOpacity style={styles.facebookButton}>
+            <Icon name="facebook" style={styles.iconFacebook} />
+            <Text style={styles.textSubmit}>{title}</Text>
+        </TouchableOpacity>
+    );
+}
+
+export default FacebookButton
+
+const styles = StyleSheet.create({
+    facebookButton: {
+        backgroundColor: COLORS.MAIN_BLUE,
+        height: 40,
+        marginTop: 25,
+        alignItems: 'center',
+        justifyContent: 'space-evenly',
+        borderWidth: 1,
+        borderColor: 'rgba(255, 255, 255, 0.07)',
+        marginBottom: 10,
+        flexDirection: 'row'
+    },
+    iconFacebook: {
+        color: COLORS.WHITE,
+        fontSize: 20
+    },
+    textSubmit: {
+        color: COLORS.WHITE,
+        fontSize: 18
+    }
+});
