+++ src/styles/app.scss
@@ -1,0 +1,39 @@
+// COLORS
+
+%white {
+    color: #ffffff
+}
+
+%main_red {
+    color: #da1705
+}
+
+%main_blue {
+    color: #0b2834
+}
+
+%grey {
+    color: #dbe9f1
+}
+
+%dark_blue {
+    color: #011d28
+}
+
+%dark_red {
+    color: #8b0303
+}
+
+%dark_grey {
+    color: #527589
+}
+
+%alt_grey {
+    color: #8b9aa1
+}
+
+%light_red {
+    color: #da807b
+}
+
+// PAGE LOGIN
