--- src/styles/constant.tsx
@@ -1,11 +1,0 @@
-export const COLORS = {
-  WHITE: "#ffffff",
-  MAIN_RED: '#da1705',
-  MAIN_BLUE: '#0b2834',
-  GREY: '#dbe9f1',
-  DARK_BLUE: '#011d28',
-  DARK_RED: '#8b0303',
-  DARK_GREY: '#527589',
-  ALT_GREY: '#8b9aa1',
-  LIGHT_RED: '#da807b'
-}
