+++ transformer.js
@@ -1,0 +1,15 @@
+var upstreamTransformer = require("metro-react-native-babel-transformer");
+var sassTransformer = require("react-native-sass-transformer");
+var postCSSTransformer = require("react-native-postcss-transformer");
+ 
+module.exports.transform = function({ src, filename, options }) {
+  if (filename.endsWith(".scss") || filename.endsWith(".sass")) {
+    return sassTransformer
+      .renderToCSS({ src, filename, options })
+      .then(css =>
+        postCSSTransformer.transform({ src: css, filename, options })
+      );
+  } else {
+    return upstreamTransformer.transform({ src, filename, options });
+  }
+};
